Aggregated remote services into their respective microservice and created a folder for the tests.
