+++ src/remote/auth-service.ts
@@ -1,0 +1,5 @@
+import { apiClient } from ".";
+
+export const userLogin = () => {
+    return apiClient.get("");
+}
