+++ src/remote/employee-service.ts
@@ -2,23 +2,22 @@ import { apiClient } from ".";
 
 // Double check these endpoints!
 
-// function to get all employees from employees microservice. 
+/**Function to get all employees from employees microservice.*/
 export const getAllEmployeesAPI = () => {
     return apiClient.get("employee/employee/employees");
 }
 
-
-// function to get employee by id
+/**Function to get employee by id*/
 export const getEmployeeByIdAPI = (id:number) => {
     return apiClient.get(`employee/employee/employees/${id}`);
 }
 
-// function to create an employee
+/**Function to create an employee*/
 export const createEmployeeAPI = (data: any) => {
     return apiClient.post('employee/employee/employees', data);
 }
 
-// function to update an existing employee
+/**Function to update an existing employee*/
 export const updateEmployeeAPI = (id:number, data: any) => {
     return apiClient.put(`employee/employee/employees/${id}`, data);
 }
