--- src/remote/login.ts
@@ -1,5 +1,0 @@
-import { loginAPI } from "../utils/APIs/loginApi";
-
-export const userLogin = () => {
-    return loginAPI.get("");
-}
