+++ src/remote/tests/campus.test.ts
@@ -1,7 +1,7 @@
-import {testRoom} from './room-details-search';
+import {getAllcampusAPI} from '../campus-service';
 
 test('Testing if data is coming back from axios request', async () => {
-    let data:any = await testRoom();
+    let data:any = await getAllcampusAPI();
     expect(data.length).not.toBe(0);
 });
