+++ src/remote/tests/employee.test.ts
@@ -1,7 +1,7 @@
-import {roomList} from './room-list-search';
+import {getAllEmployeesAPI} from '../employee-service';
+
 
 test('Testing if data is coming back from axios request', async () => {
-    let data:any = await roomList();
+    let data:any = await getAllEmployeesAPI();
     expect(data.length).not.toBe(0);
 });
-
