+++ src/remote/tests/index.test.ts
@@ -1,7 +1,7 @@
-import {userLogin} from './login';
+import {apiClient} from '../index';
 
 
 test('Testing if data is coming back from axios request',  async () => {
-    let data:any =  await userLogin();
+    let data:any =  await apiClient.get("");
     expect(data).not.toBeNull();
 });
