+++ src/remote/tests/login.test.ts
@@ -1,7 +1,7 @@
-import {apiClient} from './index';
+import {userLogin} from '../auth-service';
 
 
 test('Testing if data is coming back from axios request',  async () => {
-    let data:any =  await apiClient.get("");
+    let data:any =  await userLogin();
     expect(data).not.toBeNull();
 });
